requirements.txt
aiogram
mistralai
python-dotenv